% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ferroptosis_analysis.R
\name{prepare_sample_info}
\alias{prepare_sample_info}
\title{Prepare sample information}
\usage{
prepare_sample_info(meta_data, control_group)
}
\description{
Prepare sample information
}
\keyword{internal}
