% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ferroptosis_analysis.R
\name{save_sample_results}
\alias{save_sample_results}
\title{Save sample-level results}
\usage{
save_sample_results(result, output_dir, sample_id)
}
\description{
Save sample-level results
}
\keyword{internal}
