% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ferroptosis_analysis.R
\name{validate_inputs}
\alias{validate_inputs}
\title{Validate input data}
\usage{
validate_inputs(counts_data, meta_data, control_group, species)
}
\description{
Validate input data
}
\keyword{internal}
